Changing Bytestring to be based on lists of integers (fixes build).
After consulting with the fine folks on #pypy on freenode, it turns out
bytearrays on Rpython aren't modifiable, so we can't call .insert on
them. Python's array.array isn't available in RPython either.

However, since RPython is statically typed, a list of only integers is
just as good as bytearray.